Serialize and Deserialize Binary Tree
Two approachs to serialization. -1 denotes null nodes.
In the optimized path we don't store -1 for leaf nodes but we also
store -2 for non-leaf nodes.

This is because a full binary tree has at most n = 2^(h+1) - 1.
h is the height of the tree with the root being h = 0.
The number of leaf nodes l in the tree would be l = (n+1)/2
This means that there are always more leaf nodes than internal nodes in
the full tree scenario. In this case storing a -2 for internal nodes
requires less space than storing two -1 values for leaf nodes.